代码整理

+++ hardware/cortexM3/STM32F103/alientekV3/port/wind_type.h
@@ -66,7 +66,7 @@ typedef float w_fp32_t;
 #endif
 
 #ifndef w_fp64_t
-typedef float w_fp64_t;
+typedef double w_fp64_t;
 #endif
 
 #ifndef w_addr_t
